Rename and update ShortDeploymentNagDialog to DeploymentShortfallNagDialog
Renamed the class and associated functionalities for better clarity and alignment with its purpose. Updated the dialog text and logic to provide more detailed explanations of deployment shortfalls and guidance on addressing them. Adjusted resource keys and descriptions accordingly.
