+++ MekHQ/resources/mekhq/resources/GUI.properties
@@ -366,13 +366,16 @@ OutstandingScenariosNagDialog.text=%s, our forces are positioned for combat but 
 OutstandingScenariosNagDialog.stratCon=<br>\
   <br>We can pay 1 CVP to tactically withdraw from any scenario marked as 'Critical.' Withdrawal is\
   \ free for all other scenarios.
-PregnantCombatantNagDialog.text=%s, pregnant personnel have been identified in your TO&E, putting\
+PregnantCombatantNagDialog.text=%s, pregnant personnel have been identified in our TO&E, putting\
   \ them at risk of harm. Reassignment is advised. Do you wish to advance the day without\
   \ addressing this concern?
-
-### ShortDeploymentNagDialog Class
-ShortDeploymentNagDialog.title=Unmet Deployment Requirements
-ShortDeploymentNagDialog.text=You haven't met the deployment levels required by your contracts. Do you really wish to advance the day?
+DeploymentShortfallNagDialog.text=%s, deployment levels fall below requirements for one or more active\
+  \ contracts. While advancing the day is an option, it may strain relations with our employer. Do\
+  \ you wish to continue without addressing this?\
+  <br>\
+  <br><i>You should head to the Briefing Room and address this shortfall. Remember, Combat Teams\
+  \ assigned to the Auxiliary or Reserve roles do not count towards deployment levels. Nor do\
+  \ Combat Teams assigned to the Training role, unless the contract is Cadre Duty.</i>
 
 ### UnmaintainedUnitsNagDialog Class
 UnmaintainedUnitsNagDialog.title=Unmaintained Units
